+++ aidial_adapter_openai/chat_completions/vllm/__init__.py
@@ -1,0 +1,2 @@
+from .vllm_completion import chat_completion, extract_reasoning
+from .vllm_tokenizer import VllmTokenizer
