+++ aidial_adapter_openai/chat_completions/vllm/vllm_tokenizer.py
